High precision time offset for remapped shaders; requires game module support
Game module does timeOffset int -> float conversion, losing precision
at high server time. It can be fixed for example by using following
format in g_util.c:BuildShaderStateConfg:

	Com_sprintf(out, sizeof(out), "%s=%s:%d.%03d@",
		remappedShaders[i].oldShader,
		remappedShaders[i].newShader,
		remappedShaders[i].timeOffset / 1000,
		remappedShaders[i].timeOffset % 1000);

+++ src/renderer/tr_local.h
@@ -481,7 +481,7 @@ Ghoul2 Insert End
 	void		(*optimalStageIteratorFunc)( void );
 
 	double		clampTime;                                  // time this shader is clamped to
-	float		timeOffset;                                 // current time offset for this shader
+	double		timeOffset;                                 // current time offset for this shader
 
 	int numStates;                                    // if non-zero this is a state shader
 	struct shader_s *currentShader;                   // current state if this is a state shader
